conficker (from the intro meeting)/requirements.txt
docopt>=0.6.2
python-nmap>=0.6.0